% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dkmodel}
\alias{dkmodel}
\title{dkmodel}
\source{
\url{https://github.com/nialsiG/A-comparison-of-relief-estimates-used-in-3d-dental-topography}
}
\usage{
dkmodel
}
\description{
A list containing theoretical model surfaces corresponding to a flat surface, a single cusp, a shallow basin and a complex surface.
}
\examples{
Flat_surface <- dkmodel$flat
Single_cusp <- dkmodel$cusp
Shallow_basin <- dkmodel$basin
Complex_surface <- dkmodel$complex
}
