% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dkpongo}
\alias{dkpongo}
\title{dkpongo}
\source{
\url{https://www.morphosource.org/Detail/MediaDetail/Show/media_id/42357}
}
\usage{
dkpongo
}
\description{
A dataset containing the OES and the EDJ surfaces of a \emph{Pongo pygmaeus} upper second molar (SMF-1117)
}
\examples{
Pongo_OES <- dkpongo$OES
Pongo_EDJ <- dkpongo$EDJ
}
