% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dksetview}
\alias{dksetview}
\title{preset orientations}
\usage{
dksetview(orient = "occlusal")
}
\arguments{
\item{orient}{a character string indicating the targeted orientation (default is occlusal)}
}
\value{
sets the orientation of the 'rgl' window.
}
\description{
A function to orient 3d topographical maps using preset values.
}
\examples{
dkmap(dkmodel$cusp, inclin(dkmodel$cusp), col = "inclin", min.range = 0, max.range = 180)
dksetview()
#possible orientations are "distal", "left", "occlusal", "mesial" and "right"
}
\seealso{
\code{\link{dkmap}}
}
