% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{elev}
\alias{elev}
\title{elevation}
\usage{
elev(mesh, origin = TRUE)
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{origin}{logical, if TRUE the z of the mesh is adjusted so that the lowest z = 0
(see \code{\link{dkorigin}})}
}
\value{
A numeric vector of elevation values for all the polygons of the mesh.
}
\description{
Compute the elevation (z component of triangle barycenter).
}
\examples{
pongo_elev <- elev(dkpongo$OES)
summary(pongo_elev)

#render on a map:
dkmap(dkpongo$OES, elev(dkpongo$OES))
}
\seealso{
\code{\link{inclin}}

\code{\link{rfi}}

\code{\link{slope}}
}
