% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{orient}
\alias{orient}
\title{orientation of polygons}
\usage{
orient(mesh)
}
\arguments{
\item{mesh}{object of class mesh3d}
}
\value{
A numeric vector of occlusal orientation values in degrees for all
the polygons of the mesh.
Let the orientation from above be depicted as a trigonomical circle,
then for a tooth positioned as in Guy et al. (2015) an orientation of 0 (mesial) would
be located at an angle of pi/2, and an orientation of 90° (buccal) would
be located at an angle of 2*pi.
}
\description{
Returns the occlusal orientation (exposure in GIS)
}
\examples{
pongo_orient <- orient(dkpongo$OES)
summary(pongo_orient)

#render on a map:
#2 bins
dkmap(dkpongo$OES, orient(dkpongo$OES), col.levels = 2, col = c("yellow1", "red1"),
min.range = 0, max.range = 360, legend = TRUE, legend.type = "pie")
#4 bins
dkmap(dkpongo$OES, orient(dkpongo$OES), col.levels = 4, col = c("royalblue1",
"yellow1", "red1", "green3"), min.range = 0, max.range = 360, legend = TRUE,
legend.type = "pie")
#8 bins
dkmap(dkpongo$OES, orient(dkpongo$OES), col.levels = 8, col = "orient",
min.range = 0, max.range = 360, legend = TRUE, legend.type = "pie")
}
\seealso{
\code{\link{opc}}

\code{\link{opcr}}
}
