% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.networks.R
\docType{class}
\name{polygon.network-class}
\alias{polygon.network-class}
\alias{polygon.network}
\title{S4 class for polygon networks}
\description{
Polygon networks are subgraphs made of polygons
(i) sharing topographic features and
(ii) in contact with the rest of the subgraph by at least 1 polygon edge.
Objects of S4 class \code{polygon.network} are typically made using the function
\code{\link{poly.network}}
}
\section{Slots}{

\describe{
\item{\code{membership}}{a vector of numeric values specifying, for each triangle, the
index number of the patch to which the triangle is assigned}

\item{\code{faces}}{a vector of numeric values indicating the mesh triangle indexes}
}}

