% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{rrate}
\alias{rrate}
\title{relief rate}
\usage{
rrate(uncropped, cropped, origin = TRUE)
}
\arguments{
\item{uncropped}{object of class mesh3d. Should entirely contain the 'cropped' argument.}

\item{cropped}{object of class mesh3d. Should be part of the 'uncropped' argument.}

\item{origin}{logical, if TRUE both cropped and uncropped mesh are translated along the z-axis
so that the lowest z of the uncropped mesh = 0; see \code{\link{dkorigin}}}
}
\value{
A single relief rate value.
}
\description{
Compute the relief rate from a sub-sample of a 3d triangle mesh.
For instance, the relief rate could be computed from the portion of a molar above
the lowermost point of its central basin, compared to the whole tooth.
}
\examples{
pongo_medelev <- median(elev(dkpongo$OES))
pongo_basins <- dkcrop(dkpongo$OES, which(elev(dkpongo$OES) < pongo_medelev))
pongo_cusps <- dkcrop(dkpongo$OES, which(elev(dkpongo$OES) > pongo_medelev))

#Not run:
dkmap(pongo_basins, elev(pongo_basins))
dkmap(pongo_cusps, elev(pongo_cusps))

rrate(dkpongo$OES, pongo_basins)
rrate(dkpongo$OES, pongo_cusps)
}
\references{
\href{https://onlinelibrary.wiley.com/doi/abs/10.1002/ajpa.23916}{Thiery et al. (2019)}
}
