% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{shape.index}
\alias{shape.index}
\title{shape.index}
\usage{
shape.index(mesh, origin = TRUE)
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{origin}{logical, if TRUE the z of the mesh is adjusted so that the lowest z = 0;
see \code{\link{dkorigin}}}
}
\value{
A list of indices:
\itemize{
  \item First item
  \item Second item
  \item Second item
  \item Second item
}
}
\description{
Compute various shape indices.
}
\details{
A handful of indices have been developed to characterize the shape of natural
landscapes, including drainage basins. While some indices are very scale-sensitive (e.g.,
Gravelius' compactness coefficient, see Bendjoudi & Hubert, 2002), others are dimensionless.
Horton (1932) introduced a form factor computed as the quotient of the basin's surface area
over the square of the maximum basin length. Schumm (1956) developed a basin elongation index
computed as the quotient of twice the square root of surface area over the product of basin
length and the squareroot of pi. Lastly, Chorley et al. (1957) developed a lemniscate ratio
which corresponds to the ratio between the surface of a lemniscate of same length over the
basin area,and computed as (pi*(Length^2))/(4*Area).
}
\examples{
shape.index(dkpongo$OES)
}
\references{
\href{https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/TR013i001p00350}{Horton (1932)}

\href{https://pubs.geoscienceworld.org/gsa/gsabulletin/article-abstract/67/5/597/4811/EVOLUTION-OF-DRAINAGE-SYSTEMS-AND-SLOPES-IN?redirectedFrom=fulltext}{Schumm (1956)}

\href{http://www.ajsonline.org/content/255/2/138.extract}{Chorley et al. (1957)}

\href{http://hydrologie.org/hsj/470/hysj_47_06_0921.pdf}{Bendjoudi & Hubert (2002)}
}
