% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt.R
\name{tilt}
\alias{tilt}
\title{Occlusal tilting
Account for intra-observer error in orienting the occlusal plane by tilting a mesh along a specified axis.}
\usage{
tilt(
  mesh,
  fun,
  axis = "x",
  bins = 10,
  rotation = 30,
  randomize = FALSE,
  times = 1000,
  ...
)
}
\arguments{
\item{mesh}{object of class mesh3d}

\item{fun}{a function to apply after each occlusal tilt; the results of this function should be a single numeric value}

\item{axis}{the axis around which to rotate, either "x", "y" or "z"}

\item{bins}{the number of non-random rotations (only if randomize = FALSE)}

\item{rotation}{the maximum tilting range in degrees; tilting will be restricted between rotation and -rotation}

\item{randomize}{if TRUE randomly tilts within the rotation range}

\item{times}{the number of random rotations to perform (only if randomize = TRUE)}

\item{doMap}{boolean, if TRUE saves a snapshot of a topographic map of the slope for each rotation}
}
\value{
From the values calculated after every tilting of the occlusal plane, makes a list with the mean, geometric mean, harmonic mean, variance and standard deviation.
}
\description{
Occlusal tilting
Account for intra-observer error in orienting the occlusal plane by tilting a mesh along a specified axis.
}
\examples{
mean.slope <- function(mesh) return(mean(slope(mesh)))
tilted_slope <- tilt(dkmodel$complex, mean.slope, axis = "x", rotation = 30, times = 100, randomize = FALSE, doMap = FALSE)
}
