% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{tooth_topography}
\alias{tooth_topography}
\title{tooth_topography}
\usage{
tooth_topography(mesh, functions)
}
\arguments{
\item{mesh}{An object of class mesh3d}

\item{functions}{A list of functions to apply; these functions should accept a 'mesh' argument and return as many values as the 'mesh' face count}
}
\value{
A data frame.
}
\description{
Computes a range of dental topography variables for all triangles of a single mesh.
}
\examples{
## Prepare mesh
mesh_right <- doolkit::dkpongo$OES
## Prepare functions
Slope <- function(mesh) return(doolkit::slope(mesh))
Rotation <- function(mesh) return(doolkit::orient(mesh))
fun_list <- list("slope" = Slope, "rotation" = Rotation)
## Batch tooth analysis
tooth_topo_dataframe <- doolkit::tooth_topography(mesh_right, fun_list)
}
