% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{tridist}
\alias{tridist}
\title{Distance between two paired meshes}
\usage{
tridist(origin, target)
}
\arguments{
\item{origin}{object of class mesh3d; should be the outer enamel surface}

\item{target}{object of class mesh3d; should be the enamel-dentine junction}
}
\description{
Compute distance from an origin mesh to a target mesh with paired triangle indices.
A mesh can be paired to another using the 'tridx' method.
}
\examples{
""
}
\seealso{
\code{\link{tridx}}
}
