% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch.multi}
\alias{batch.multi}
\title{batch.multi}
\usage{
batch.multi(files, functions, do.parallel = TRUE)
}
\arguments{
\item{files}{A list of surface files; accepted formats are as in RVCG::vcgImport}

\item{functions}{A list of functions to apply; these functions should accept a 'mesh' argument and return a single value}
}
\value{
A data frame.
}
\description{
Computes a range of dental topography variables for a range of meshes.
}
\examples{
## Prepare a list of files
meshes_right <- list("pongo_OES" = doolkit::dkpongo$OES, "pongo_EDJ" = doolkit::dkpongo$EDJ)
## Prepare functions
Mss <- function(mesh) return(mean(doolkit::slope(mesh)))
Rfi <- function(mesh) return(doolkit::rfi(mesh, method = "Boyer"))
Opcr <- function(mesh) return(doolkit::opcr(mesh)$opcr)
fun_list <- list("mss" = Mss, "rfi" = Rfi, "opcr" = Opcr)
## Batch sample analysis
sample_topo_dataframe <- doolkit::sample_topography(meshes_right, fun_list)
}
\seealso{
\code{\link[Rvcg]{vcgImport}}
}
