% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch.single}
\alias{batch.single}
\title{batch.single}
\usage{
batch.single(mesh, functions)
}
\arguments{
\item{mesh}{An object of class mesh3d}

\item{functions}{A list of functions to apply; these functions should accept a 'mesh' argument and return as many values as the 'mesh' face count}
}
\value{
A data frame.
}
\description{
Computes a range of dental topography variables for all triangles of a single mesh.
}
\examples{
## Prepare mesh
my_mesh <- doolkit::dkmodel$cusp
## Prepare functions
Slope <- function(mesh) return(doolkit::slope(mesh))
Rotation <- function(mesh) return(doolkit::orient(mesh))
fun_list <- list("slope" = Slope, "rotation" = Rotation)
## Single mesh batch analysis
single_mesh_topography <- doolkit::batch.single(my_mesh, fun_list)
}
