% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dkborder}
\alias{dkborder}
\title{dkborder}
\usage{
dkborder(mesh)
}
\arguments{
\item{mesh}{object of class mesh3d}
}
\value{
A vector of indices corresponding to the triangles with at least one vertex on the border of the mesh.
}
\description{
Selects the border of a 3d triangle mesh.
}
\examples{
pongo_border <- dkborder(dkpongo$OES)

# Map the border in orange:
is_border <- rep(1, Rvcg::nfaces(dkpongo$OES))
is_border[pongo_border] <- 2
dkmap(dkpongo$OES, is_border, col = c("white", "#E69F00"), col.levels = 2, legend = FALSE,
scalebar = FALSE)

# Compare with vcgBorder from the R package Rvcg, in blue:
pongo_vcgborder <- which(Rvcg::vcgBorder(dkpongo$OES)$borderit == TRUE)
is_border[pongo_vcgborder] <- 3
dkmap(dkpongo$OES, is_border, col = c("white", "#E69F00", "#56B4E9"), col.levels = 3,
legend = FALSE, scalebar = FALSE)
#As you can see, it all depends on what you want to select!

}
