% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkmodel}
\alias{dkmodel}
\title{dkmodel}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{https://github.com/nialsiG/A-comparison-of-relief-estimates-used-in-3d-dental-topography}
}
\usage{
dkmodel
}
\description{
A list containing theoretical model surfaces corresponding to a flat surface, a single cusp, a shallow basin and a complex surface.
}
\keyword{datasets}
