% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{dksave}
\alias{dksave}
\title{dksave}
\usage{
dksave(df, file, format = ".txt")
}
\arguments{
\item{df}{An object of class data.frame}

\item{file}{The path to which the data.frame must be saved}

\item{format}{The format to which the data frame should be saved, either ".txt", ".csv", ".RDS" or ".xslx" (default is "txt"))}
}
\value{
A 'format' file saved at the 'path' argument.
}
\description{
Save a data frame.
}
\examples{
# Build a topographic data frame
mesh_right <- doolkit::dkpongo$OES
Slope <- function(mesh) return(doolkit::slope(mesh))
Rotation <- function(mesh) return(doolkit::orient(mesh))
fun_list <- list("slope" = Slope, "rotation" = Rotation)
tooth_topo_dataframe <- doolkit::tooth_topography(mesh_right, fun_list)
#
# Save the dataframe to file...
# ...as .txt:
tmpfile <- tempfile("topographic_table", fileext = ".txt")
dksave(tooth_topo_dataframe, tmpfile, format = ".txt")
# ...as .csv:
tmpfile <- tempfile("topographic_table", fileext = ".csv")
dksave(tooth_topo_dataframe, tmpfile, format = ".csv")
# ...as .RDS:
tmpfile <- tempfile("topographic_table", fileext = ".RDS")
dksave(tooth_topo_dataframe, tmpfile, format = ".RDS")
# ...as .xslx:
tmpfile <- tempfile("topographic_table", fileext = ".xslx")
dksave(tooth_topo_dataframe, tmpfile, format = ".xslx")

unlink(tmpfile)
}
