% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{sample_topography}
\alias{sample_topography}
\title{sample_topography}
\usage{
sample_topography(meshes, functions)
}
\arguments{
\item{meshes}{A list of objects of class mesh3d}

\item{functions}{A list of functions to apply}
}
\value{
A data frame.
}
\description{
Computes a range of dental topography variables for a range of meshes.
}
\examples{
#
}
