% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{slope}
\alias{slope}
\title{slope}
\usage{
slope(mesh)
}
\arguments{
\item{mesh}{object of class mesh3d}
}
\value{
A numeric vector of slope values for all the polygons of the mesh.
}
\description{
Compute slope i.e. the angle between triangles and the horizontal plane in degrees, comprised
between 0 and 90.
}
\examples{
pongo_slope <- slope(dkpongo$OES)
summary(pongo_slope)

#render on a map:
dkmap(dkpongo$OES, slope(dkpongo$OES), col.levels = 9, col = "slope",
min.range = 0, max.range = 90, legend = TRUE)
}
\seealso{
\code{\link{inclin}}
}
