% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{tooth_topography}
\alias{tooth_topography}
\title{tooth_topography}
\usage{
tooth_topography(mesh, functions)
}
\arguments{
\item{mesh}{An object of class mesh3d}

\item{functions}{A list of functions to apply}
}
\value{
A data frame.
}
\description{
Computes a range of dental topography variables for all triangles of a single mesh.
}
\examples{
#
}
