% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridx.R
\name{tridx}
\alias{tridx}
\title{tridx}
\usage{
tridx(oes, edj, ray = FALSE)
}
\arguments{
\item{oes}{object of class mesh3d; should be the outer enamel surface}

\item{edj}{object of class mesh3d; should be the enamel-dentine junction}

\item{ray}{logical, if TRUE the search is along vertex normals (default is FALSE)}
}
\value{
A data frame.
}
\description{
Get the pairwise indexing of a 3D mesh relative to another.
}
\examples{
tridx <- data.frame()

}
