% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridx.R
\name{tridx}
\alias{tridx}
\title{tridx}
\usage{
tridx(oes, edj, method = "nearest", direction = -1, epsilon = 1e-07)
}
\arguments{
\item{oes}{object of class mesh3d; should be the outer enamel surface}

\item{edj}{object of class mesh3d; should be the enamel-dentine junction}

\item{method}{String indicating which method should be used to find triangle pairs.
Default is 'nearest', which find the absolute nearest 'edj' triangle.
Alternative choices: 'normal' follows the normal to surface A, 'ortho' follows the Z-axis.}
}
\value{
A data frame.
}
\description{
Get the pairwise indexing of a 3D mesh relative to another.
}
\examples{
result <- doolkit::tridx(oes = doolkit::dkmodel$cusp, edj = doolkit::dkmodel$flat, method = "nearest")

}
